create stream-related types
Created simpleblob.NewReader and simpleblob.NewWriter functions, along with ReaderStorage and WriterStorage interface types.
Those functions allow to provide an optimized way to write to and read from backend if those interfaces are satsfied.
If an interface is not satisfied by backend, a fallback implementation is used, akin to `fs.GlobFS`.