implement ReaderStorage for S3 backend
The Object returned by MinIO client satisfies io.ReadCloser, so this implementation is very straightforward.

For writing, the MinIO client takes a reader and needs the full size of the stored object.
Thus, we'd need to exhaust the writer into a buffer, take its size and provide it to the client,
which is basically what the fallback implementation of NewWriter does.